+++ app/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.21.0](https://github.com/efiten/core-hunter/compare/app-v1.20.0...app-v1.21.0) (2026-08-28)
+
+
+### Features
+
+* **app:** name a trace reply after the node we pinged ([#483](https://github.com/efiten/core-hunter/issues/483)) ([04eded1](https://github.com/efiten/core-hunter/commit/04eded1708612b9462888e0d4925e0b4b5d285ef))
+* **web:** name the picked target on the picker button ([#499](https://github.com/efiten/core-hunter/issues/499)) ([cabe596](https://github.com/efiten/core-hunter/commit/cabe5969eec5f3038ecbe3aa68d018b8af981464))
+
 ## [1.20.0](https://github.com/efiten/core-hunter/compare/app-v1.19.1...app-v1.20.0) (2026-08-26)